Add collapsible sections to task filters
